pyproject.toml
[project]
name = "web-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.9",
    "patchright>=1.57.2",
    "playwright>=1.57.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
